+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
+- **Breaking**: Renamed `file_format` API option to `format`
 - Add `bestaudio` and `bestvideo` formats
 
 ## 3.6.5 2021 Sep 15